+++ test/analyzer/index_test.cpp
@@ -503,10 +503,10 @@ $value: 'hey')
 TEST_F(IndexTest, namedArgBuiltinMethod) {
   unsigned short modId;
   const char *content = R"E(
-"234".slice($to:444, $from: 2222)
-"2345".slice($from:345, $to: 333)
-[234].slice($to:444, $from: 2222)
-['ss'].slice($from:345, $to: 333)
+"234".slice($end:444, $start: 2222)
+"2345".slice($start:345, $end: 333)
+[234].slice($end:444, $start: 2222)
+['ss'].slice($start:345, $end: 333)
 )E";
 
   ASSERT_NO_FATAL_FAILURE(this->doAnalyze(content, modId, {.declSize = 0, .symbolSize = 12}));
@@ -515,26 +515,26 @@ TEST_F(IndexTest, namedArgBuiltinMethod) {
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = modId, .position = {.line = 1, .character = 25}},
                      {
-                         {modId, "(1:21~1:26)"}, // $from: 2222
-                         {modId, "(2:13~2:18)"}, // $from:345
+                         {modId, "(1:22~1:28)"}, // $start: 2222
+                         {modId, "(2:13~2:19)"}, // $start:345
                      }));
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = modId, .position = {.line = 2, .character = 26}},
                      {
-                         {modId, "(1:12~1:15)"}, // $to:444
-                         {modId, "(2:24~2:27)"}, // $to: 333
+                         {modId, "(1:12~1:16)"}, // $end:444
+                         {modId, "(2:25~2:29)"}, // $end: 333
                      }));
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = modId, .position = {.line = 3, .character = 25}},
                      {
-                         {modId, "(3:21~3:26)"}, // $from: 2222
-                         {modId, "(4:13~4:18)"}, // $from:345
+                         {modId, "(3:22~3:28)"}, // $start: 2222
+                         {modId, "(4:13~4:19)"}, // $start:345
                      }));
   ASSERT_NO_FATAL_FAILURE(
       this->findRefs(Request{.modId = modId, .position = {.line = 4, .character = 26}},
                      {
-                         {modId, "(3:12~3:15)"}, // $to:444
-                         {modId, "(4:24~4:27)"}, // $to: 333
+                         {modId, "(3:12~3:16)"}, // $end:444
+                         {modId, "(4:25~4:29)"}, // $end: 333
                      }));
 }
 
@@ -1731,15 +1731,16 @@ TEST_F(IndexTest, hoverBuiltin) {
                                       "```arsh\nfunction size(): Int for [Int]\n```"));
   ASSERT_NO_FATAL_FAILURE(
       this->hover("''.slice(0)", Position{.line = 0, .character = 5},
-                  "```arsh\nfunction slice(from: Int, to: Int?): String for String\n```"));
+                  "```arsh\nfunction slice(start: Int, end: Int?): String for String\n```"));
   ASSERT_NO_FATAL_FAILURE(
       this->hover("[23:(12,(34,56,))].\nclear()", Position{.line = 1, .character = 2},
                   "```arsh\nfunction clear(): Void for [Int : (Int, (Int, Int))]\n```"));
-  ASSERT_NO_FATAL_FAILURE(this->hover("'2345'.slice(\n$to:2345)", 1, "```arsh\nvar to: Int?\n```"));
   ASSERT_NO_FATAL_FAILURE(
-      this->hover("[1234].slice(\n$from:2345)", 1, "```arsh\nvar from: Int\n```"));
+      this->hover("'2345'.slice(\n$end:2345)", 1, "```arsh\nvar end: Int?\n```"));
   ASSERT_NO_FATAL_FAILURE(
-      this->hover("[1234].slice(\n$to:345,$from:2345)", 1, "```arsh\nvar to: Int?\n```"));
+      this->hover("[1234].slice(\n$start:2345)", 1, "```arsh\nvar start: Int\n```"));
+  ASSERT_NO_FATAL_FAILURE(
+      this->hover("[1234].slice(\n$end:345,$start:2345)", 1, "```arsh\nvar end: Int?\n```"));
   ASSERT_NO_FATAL_FAILURE(
       this->hover("['23':1234].remove(\n$key:'2')", 1, "```arsh\nvar key: String\n```"));
 }