+++ test/analyzer/rename_test.cpp
@@ -723,19 +723,19 @@ new Interval($bb: 0, $ee: 100).dist(
 
 TEST_F(RenameTest, namedArg2) { // for builtin method
   const char *content = R"(
-"234".slice($to:444, $from: 2222)
-"2345".slice($from:345, $to: 333)
-[234].slice($to:444, $from: 2222)
-['ss'].slice($from:345, $to: 333)
+"234".slice($end:444, $start: 2222)
+"2345".slice($start:345, $end: 333)
+[234].slice($end:444, $start: 2222)
+['ss'].slice($start:345, $end: 333)
 )";
   ASSERT_NO_FATAL_FAILURE(this->doAnalyze(content, 1));
-  ASSERT_NO_FATAL_FAILURE(this->rename(Request{.modId = 1, .line = 1, .character = 24}, "start",
+  ASSERT_NO_FATAL_FAILURE(this->rename(Request{.modId = 1, .line = 1, .character = 24}, "begin",
                                        RenameValidationStatus::BUILTIN));
-  ASSERT_NO_FATAL_FAILURE(this->rename(Request{.modId = 1, .line = 2, .character = 16}, "start",
+  ASSERT_NO_FATAL_FAILURE(this->rename(Request{.modId = 1, .line = 2, .character = 28}, "stop",
                                        RenameValidationStatus::BUILTIN));
-  ASSERT_NO_FATAL_FAILURE(this->rename(Request{.modId = 1, .line = 2, .character = 14}, "start",
+  ASSERT_NO_FATAL_FAILURE(this->rename(Request{.modId = 1, .line = 2, .character = 14}, "step",
                                        RenameValidationStatus::BUILTIN));
-  ASSERT_NO_FATAL_FAILURE(this->rename(Request{.modId = 1, .line = 4, .character = 25}, "start",
+  ASSERT_NO_FATAL_FAILURE(this->rename(Request{.modId = 1, .line = 4, .character = 18}, "begin",
                                        RenameValidationStatus::BUILTIN));
 }